associated product count with add/delete and showed it on checkout

+++ BE/addtocart/src/main/java/com/amazon/addtocart/service/AddToCartService.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import com.amazon.addtocart.entity.Cart;
 import com.amazon.addtocart.entity.CartDetails;
+import com.amazon.addtocart.entity.CartProduct;
 import com.amazon.addtocart.entity.Product;
 import com.amazon.addtocart.repository.AddToCartRepository;
 
@@ -23,11 +24,23 @@ public class AddToCartService {
 	public ProductDetailsProxy productDetailsProxy;
 
 	public void addToCart(Cart cart) {
-		addToCartRepository.save(cart);
+		if(cart.productCount > 1) {
+			Cart cartProduct = addToCartRepository.findByProductId(cart.productId);
+			cartProduct.setProductCount(cart.productCount);
+			addToCartRepository.save(cartProduct);
+		} else {
+			addToCartRepository.save(cart);
+		}
 	}
 
 	public void deleteFromCart(Cart cart) {
-		addToCartRepository.deleteByProductId(cart.productId);
+		if(cart.productCount > 0) {
+			Cart cartProduct = addToCartRepository.findByProductId(cart.productId);
+			cartProduct.setProductCount(cart.productCount);
+			addToCartRepository.save(cartProduct);
+		} else {
+			addToCartRepository.deleteByProductId(cart.productId);
+		}
 	}
 
 	public CartDetails displayAllCartItemsForUser(UUID userId) {
@@ -36,11 +49,12 @@ public class AddToCartService {
 		cartDetails.setUserId(userId);
 
 		ArrayList<Cart> cartItems = addToCartRepository.findByUserId(userId);
-		ArrayList<Product> products = new ArrayList<>();
+		ArrayList<CartProduct> products = new ArrayList<>();
 		for(Cart cartItem: cartItems) {
 			System.out.println("Cart Item: " + cartItem.getProductId());
 			Product product = productDetailsProxy.getProductById(cartItem.getProductId());
-			products.add(product);
+			int productCount = cartItem.getProductCount();
+			products.add(new CartProduct(product, productCount));
 		}
 		
 		cartDetails.setList(products);
